Add option for remote Tor on a server
The option to use Tor from a remote server allows the user to keep a clean local system but enjoy Tor operated elsewhere.  The script will use "Tor-Remote" to create an SSH connection to the server, with port forwarding. Locally, the web browser (or any Torrified apps) should be configured to use a SOCKS version 5 proxy on port 9050.